+++ scripts/popup.js
@@ -1,21 +1,46 @@
 var background_page = browser.extension.getBackgroundPage();
 
 window.onload = function() {
+	localizeHtmlPage();
 	browser.storage.local.get().then( (option) => {
-		let last_message = option.last_message || "No startup sync done yet.";
-		document.getElementById("popupMessage").appendChild( document.createTextNode(last_message) );
+		let last_message = option.last_message || browser.i18n.getMessage("popupNoSync");
+		document.getElementById("popupMessage").appendChild(document.createTextNode(last_message));
 	});
+	
+	document.getElementById("export").addEventListener("click", manualExport);
+	document.getElementById("import").addEventListener("click", manualImport);
+	document.getElementById("remove").addEventListener("click", manualRemove);
+	document.getElementById("settings").addEventListener("click", background_page.openSettings);
 };
 
+function localizeHtmlPage() {
+    var objects = document.getElementsByTagName('html');
+    for (var j = 0; j < objects.length; j++)
+    {
+        var obj = objects[j];
+
+        var valStrH = obj.innerHTML.toString();
+        var valNewH = valStrH.replace(/__MSG_(\w+)__/g, function(match, v1)
+        {
+            return v1 ? browser.i18n.getMessage(v1) : "";
+        });
+
+        if(valNewH != valStrH)
+        {
+            obj.innerHTML = valNewH;
+        }
+    }
+}
+
 function manualRemove() {
-	if(confirm("All current bookmarks are removed. Are you sure?")) {
+	if(confirm(browser.i18n.getMessage("optionsBTNRemoveHint"))) {
 		background_page.removeAllMarks();
 		browser.storage.local.set({last_s: 1});
 	}
 }
 
 function manualImport() {
-	if(confirm("The bookmarks stored on the server are now added. Are you sure?")) {
+	if(confirm(browser.i18n.getMessage("popupImportConfirm"))) {
 		background_page.checkSettings();
 		
 		if(background_page.s_type == 'PHP') {
@@ -38,7 +63,3 @@ function manualExport() {
 	}
 }
 
-document.getElementById("export").addEventListener("click", manualExport);
-document.getElementById("import").addEventListener("click", manualImport);
-document.getElementById("remove").addEventListener("click", manualRemove);
-document.getElementById("settings").addEventListener("click", background_page.openSettings);